Move update-order-methods from SortProcedureServ to ScannerProcedureServ
 - To accommodate cancellations
